liveness_blink_detection/requirements.txt
cmake==3.22.2
dlib==19.23.0
easydict==1.9
filelock==3.12.3
imutils==0.5.4
Jinja2==3.1.2
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
nptyping==1.4.3
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
opencv-python==4.8.0.76
pyexifinfo==0.4.0
PyExifTool==0.5.5
scipy==1.10.1
sympy==1.12
torch==2.0.1
triton==2.0.0
typing-extensions==4.7.1
typish==1.9.3
